pyproject.toml
[project]
name = "langCracked"
version = "0.1.0"
description = "Langchain and Langgraph basics for dummies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.13",
    "langchain>=0.3.26",
    "langchain-aws>=0.2.27",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
